:art: style(dto): move dto into category with request and response
